Add files via upload
Baseline Watermelon is the main code. White line black dot is the plugin for displaying the white line and controlling the rotation. Random Image storage is the plugin that displays a random image (with replacement so far) at a random orientation. 
The rest are images or just other default jspsych plugins.